requirements.txt
pyrogram
pyKrishna
tgcrypto
python-dotenv[cli]
python-dotenv
gitpython
psutil
py-cpuinfo 
telegraph
